Dell P2715Q DisplayPort 1.2 and MST settings
I discovered your page here when I searched on github and your page came up as the first hit! :) Am troubleshooting the setup of my new P2715Q at home and ran into problems with my mid-2013 MacBook Air not getting the max resolution (when it should be able to).